restore settings page
